pyproject.toml
[project]
name = "agentic-zerodha-platform"
version = "0.1.0"
description = "Agentic Zerodha platform with human-in-loop approvals"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "Django>=5.1,<6.0",
  "djangorestframework>=3.15,<4.0",
  "django-environ>=0.11,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "celery[redis]>=5.4,<6.0",
  "gunicorn>=22.0,<24.0",
  "kiteconnect>=5.0,<6.0",
  "cryptography>=44.0,<45.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-django>=4.11,<5.0",
  "ruff>=0.9,<1.0",
  "mypy>=1.14,<2.0",
  "types-redis>=4.6.0.20241004",
]

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/config", "src/apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "DJ"]
ignore = ["DJ008"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = []
exclude = "(^|/)migrations/"
